+++ src/main/java/ewing/user/UserService.java
@@ -5,19 +5,21 @@ import ewing.dao.entity.User;
 import ewing.user.vo.FindUserParam;
 import ewing.user.vo.UserWithRole;
 
+import java.math.BigInteger;
+
 /**
  * 用户服务接口。
  **/
 public interface UserService extends UserBeans {
 
-    User getUser(Long userId);
+    User getUser(BigInteger userId);
 
-    Long addUserWithRole(UserWithRole userWithRole);
+    BigInteger addUserWithRole(UserWithRole userWithRole);
 
     long updateUserWithRole(UserWithRole userWithRole);
 
     Page<UserWithRole> findUserWithRole(FindUserParam findUserParam);
 
-    long deleteUser(Long userId);
+    long deleteUser(BigInteger userId);
 
 }
